Moved MetadataStore to a separate module
